Updated DuneScaling to work with Chelsea's Data.py
Additional functions to measure the time, position and height of the dune using point masses. Additional code to put a linear regression into the graphs. Added the line for the 38 degree angle of repose for snow dunes. Made a function to find any errors in the heights in the .csv file. Added the perspective correction from the video on 1/24/2018 to the code.